# apps/web/src/api/openapi.json
{
  "components": {
    "schemas": {
      "Ack": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "ok": {
            "default": true,
            "title": "Ok",
            "type": "boolean"
          }
        },
        "title": "Ack",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OptimizeRequest": {
        "additionalProperties": false,
        "properties": {
          "candidate_stores": {
            "default": 10,
            "description": "N nearest stores the subsets are drawn from",
            "maximum": 15.0,
            "minimum": 1.0,
            "title": "Candidate Stores",
            "type": "integer"
          },
          "clubs": {
            "items": {
              "type": "string"
            },
            "title": "Clubs",
            "type": "array"
          },
          "home_store_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "The user's usual store; savings are measured against it only",
            "title": "Home Store Id"
          },
          "include_online": {
            "default": false,
            "title": "Include Online",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/BasketItem"
            },
            "maxItems": 200,
            "minItems": 1,
            "title": "Items",
            "type": "array"
          },
          "location": {
            "$ref": "#/components/schemas/Location"
          },
          "max_stores": {
            "default": 2,
            "maximum": 3.0,
            "minimum": 1.0,
            "title": "Max Stores",
            "type": "integer"
          },
          "min_split_saving": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "default": "25",
            "description": "A split is recommended only above this net saving",
            "title": "Min Split Saving"
          },
          "solver": {
            "default": "heuristic",
            "description": "milp handles cross-item promos and quantity rounding (phase 2)",
            "enum": [
              "heuristic",
              "milp"
            ],
            "title": "Solver",
            "type": "string"
          },
          "travel": {
            "$ref": "#/components/schemas/TravelSettings"
          }
        },
        "required": [
          "items",
          "location"
        ],
        "title": "OptimizeRequest",
        "type": "object"
      },
      "OptimizeResponse": {
        "additionalProperties": false,
        "properties": {
          "disclaimer_he": {
            "default": "המחיר הקובע הוא בקופה.",
            "title": "Disclaimer He",
            "type": "string"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "minimum_effort": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Plan"
              },
              {
                "type": "null"
              }
            ],
            "description": "The home store itself; null without a home store"
          },
          "single": {
            "$ref": "#/components/schemas/Plan"
          },
          "solver": {
            "default": "heuristic",
            "enum": [
              "heuristic",
              "milp"
            ],
            "title": "Solver",
            "type": "string"
          },
          "split": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Plan"
              },
              {
                "type": "null"
              }
            ],
            "description": "null when no split beats min_split_saving"
          },
          "subsets_evaluated": {
            "title": "Subsets Evaluated",
            "type": "integer"
          }
        },
        "required": [
          "single",
          "subsets_evaluated",
          "generated_at"
        ],
        "title": "OptimizeResponse",
        "type": "object"
      },
...
      "Profile": {
        "additionalProperties": false,
        "properties": {
          "clubs": {
            "items": {
              "type": "string"
            },
            "title": "Clubs",
            "type": "array"
          },
          "consent_location": {
            "default": false,
            "description": "Required to store a neighborhood location",
            "title": "Consent Location",
            "type": "boolean"
          },
          "cost_per_km": {
            "default": "1.2",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Cost Per Km",
            "type": "string"
          },
          "diet_flags": {
            "items": {
              "type": "string"
            },
            "title": "Diet Flags",
            "type": "array"
          },
          "exists": {
            "description": "false when nothing is stored yet and these are the defaults",
            "title": "Exists",
            "type": "boolean"
          },
          "extra_stop_value": {
            "default": "25",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Extra Stop Value",
            "type": "string"
          },
          "flex_defaults": {
            "additionalProperties": {
              "enum": [
                "exact",
                "any_brand",
                "close"
              ],
              "type": "string"
            },
            "title": "Flex Defaults",
            "type": "object"
          },
          "home_store_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Home Store Id"
          },
          "kosher_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kosher Level"
          },
          "max_stores": {
            "default": 2,
            "maximum": 3.0,
            "minimum": 1.0,
            "title": "Max Stores",
            "type": "integer"
          },
          "neighborhood_lat": {
            "anyOf": [
              {
                "maximum": 90.0,
                "minimum": -90.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Stored rounded to 3 decimals (about 100 m)",
            "title": "Neighborhood Lat"
          },
          "neighborhood_lon": {
            "anyOf": [
              {
                "maximum": 180.0,
                "minimum": -180.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Neighborhood Lon"
          },
          "radius_m": {
            "default": 5000,
            "maximum": 15000.0,
            "minimum": 500.0,
            "title": "Radius M",
            "type": "integer"
          },
          "theme": {
            "default": "system",
            "enum": [
              "system",
              "light",
              "dark"
            ],
            "title": "Theme",
            "type": "string"
          },
          "travel_mode": {
            "default": "car",
            "enum": [
              "car",
              "walk_transit",
              "delivery"
            ],
            "title": "Travel Mode",
            "type": "string"
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "exists"
        ],
        "title": "Profile",
        "type": "object"
      },
      "ProfileUpdate": {
        "additionalProperties": false,
        "properties": {
          "clubs": {
            "items": {
              "type": "string"
            },
            "title": "Clubs",
            "type": "array"
          },
          "consent_location": {
            "default": false,
            "description": "Required to store a neighborhood location",
            "title": "Consent Location",
            "type": "boolean"
          },
          "cost_per_km": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "default": "1.2",
            "title": "Cost Per Km"
          },
          "diet_flags": {
            "items": {
              "type": "string"
            },
            "title": "Diet Flags",
            "type": "array"
          },
          "extra_stop_value": {
            "anyOf": [
              {
                "maximum": 50.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "default": "25",
            "title": "Extra Stop Value"
          },
          "flex_defaults": {
            "additionalProperties": {
              "enum": [
                "exact",
                "any_brand",
                "close"
              ],
              "type": "string"
            },
            "title": "Flex Defaults",
            "type": "object"
          },
          "home_store_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Home Store Id"
          },
          "kosher_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kosher Level"
          },
          "max_stores": {
            "default": 2,
            "maximum": 3.0,
            "minimum": 1.0,
            "title": "Max Stores",
            "type": "integer"
          },
          "neighborhood_lat": {
            "anyOf": [
              {
                "maximum": 90.0,
                "minimum": -90.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Stored rounded to 3 decimals (about 100 m)",
            "title": "Neighborhood Lat"
          },
          "neighborhood_lon": {
            "anyOf": [
              {
                "maximum": 180.0,
                "minimum": -180.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Neighborhood Lon"
          },
          "radius_m": {
            "default": 5000,
            "maximum": 15000.0,
            "minimum": 500.0,
            "title": "Radius M",
            "type": "integer"
          },
          "theme": {
            "default": "system",
            "enum": [
              "system",
              "light",
              "dark"
            ],
            "title": "Theme",
            "type": "string"
          },
          "travel_mode": {
            "default": "car",
            "enum": [
              "car",
              "walk_transit",
              "delivery"
            ],
            "title": "Travel Mode",
            "type": "string"
          }
        },
        "title": "ProfileUpdate",
        "type": "object"
      },
...
      "ShoppingList": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_recurring": {
            "title": "Is Recurring",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ListItem"
            },
            "title": "Items",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "is_recurring",
          "items",
          "created_at",
          "updated_at"
        ],
        "title": "ShoppingList",
        "type": "object"
      },
      "ShoppingListIn": {
        "additionalProperties": false,
        "properties": {
          "is_recurring": {
            "default": false,
            "title": "Is Recurring",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ListItemIn"
            },
            "maxItems": 200,
            "title": "Items",
            "type": "array"
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "ShoppingListIn",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Semantic supermarket price comparison for Israel. Prices are from the chains' transparency files; המחיר הקובע הוא בקופה.",
    "title": "SmartCart API",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/me": {
      "delete": {
        "operationId": "delete_me_me_delete",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ack"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Me",
        "tags": [
          "me"
        ]
      }
    },
...
    "/me/lists": {
      "get": {
        "operationId": "list_lists_me_lists_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ShoppingList"
                  },
                  "title": "Response List Lists Me Lists Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Lists",
        "tags": [
          "me"
        ]
      },
      "post": {
        "operationId": "create_list_me_lists_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShoppingListIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShoppingList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create List",
        "tags": [
          "me"
        ]
      }
    },
    "/me/lists/{list_id}": {
      "delete": {
        "operationId": "delete_list_me_lists__list_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "list_id",
            "required": true,
            "schema": {
              "title": "List Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete List",
        "tags": [
          "me"
        ]
      },
      "get": {
        "operationId": "get_list_me_lists__list_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "list_id",
            "required": true,
            "schema": {
              "title": "List Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShoppingList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get List",
        "tags": [
          "me"
        ]
      },
      "put": {
        "description": "Replace the list's name, recurrence and items.",
        "operationId": "update_list_me_lists__list_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "list_id",
            "required": true,
            "schema": {
              "title": "List Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShoppingListIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShoppingList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update List",
        "tags": [
          "me"
        ]
      }
    },
...
    "/me/profile": {
      "get": {
        "operationId": "get_profile_me_profile_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Profile"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Profile",
        "tags": [
          "me"
        ]
      },
      "put": {
        "operationId": "put_profile_me_profile_put",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Profile"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Profile",
        "tags": [
          "me"
        ]
      }
    },
...
    "/optimize": {
      "post": {
        "operationId": "optimize_optimize_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OptimizeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OptimizeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Optimize",
        "tags": [
          "basket"
        ]
      }
    },
    "/optimize/swaps": {
      "post": {
        "operationId": "swap_suggestions_optimize_swaps_post",
        "parameters": [
          {
            "in": "query",
            "name": "store_id",
            "required": true,
            "schema": {
              "title": "Store Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompareRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SwapSuggestionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Swap Suggestions",
        "tags": [
          "basket"
        ]
      }
    },
...
  }
}